+++ src/cmd/internal/goobj/read.go
