+++ src/cmd/internal/goobj/read_test.go
